[SOL] rework test harness codegen a bit
This change reworks the way we generate the test harness, it removes  most
bpf-only changes,  implements a custom start entry for the bpf arch, and tweaks
the link stage to set the correct entry point depending on whether we're
compiling a library or tests.

With this change, artifacts produced by rustc --test can be loaded in rbpf like
any other library.